# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Cedar & Stone chalet reservations and admin operations
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Chalet {
  id: number;
  title: string;
  slug: string;
  description: string;
  capacity: number;
  weekdayPrice: number;
  weekendPrice: number;
  latitude: number;
  longitude: number;
  mainImageUrl: string;
  extraImageUrls: string[];
  createdAt: string;
}

export interface ChaletInput {
  /** @minLength 2 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @minimum 1 */
  capacity: number;
  /** @minimum 0 */
  weekdayPrice: number;
  /** @minimum 0 */
  weekendPrice: number;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
  /** @minLength 1 */
  mainImageUrl: string;
  extraImageUrls: string[];
}

export type ChaletUpdate = ChaletInput;

export interface Availability {
  blockedDates: string[];
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  Pending: 'Pending',
  Accepted: 'Accepted',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface Booking {
  id: number;
  chaletId: number;
  chaletTitle: string;
  guestName: string;
  phone: string;
  checkIn: string;
  checkOut: string;
  totalPrice: number;
  status: BookingStatus;
  createdAt: string;
}

export interface BookingInput {
  /** @minimum 1 */
  chaletId: number;
  /** @minLength 2 */
  guestName: string;
  /** @minLength 5 */
  phone: string;
  checkIn: string;
  checkOut: string;
}

export type BookingStatusInputStatus = typeof BookingStatusInputStatus[keyof typeof BookingStatusInputStatus];


export const BookingStatusInputStatus = {
  Pending: 'Pending',
  Accepted: 'Accepted',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface BookingStatusInput {
  status: BookingStatusInputStatus;
}

export interface Admin {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AdminInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface PasswordInput {
  /** @minLength 8 */
  password: string;
}

export interface Dashboard {
  revenue: number;
  pendingCount: number;
  acceptedCount: number;
  completedCount: number;
  cancelledCount: number;
  chaletCount: number;
  recentBookings: Booking[];
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  /** @pattern ^image/ */
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadRequest;
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = void;

/**
 * Resource not found
 */
export type NotFoundResponse = void;