# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Cedar & Stone chalet reservations and admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  name: string;
  email: string;
  createdAt: Date;
}