# lib/api-zod/src/generated/types/adminInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Cedar & Stone chalet reservations and admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface AdminInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}